requirements.txt
absl-py==0.15.0
autograd==1.3
cachetools==4.2.4
certifi==2021.10.8
charset-normalizer==2.0.7
cycler==0.10.0
flatbuffers==2.0
future==0.18.2
google-auth==2.3.0
google-auth-oauthlib==0.4.6
grpcio==1.41.0
idna==3.3
imageio==2.9.0
jax==0.2.24
jaxlib==0.1.73
joblib==1.1.0
KDEpy==1.1.0
kiwisolver==1.3.2
Markdown==3.3.4
matplotlib==3.4.3
netcal==1.2.1
networkx==2.6.3
numpy==1.21.3
oauthlib==3.1.1
opt-einsum==3.3.0
pandas==1.3.4
patsy==0.5.2
Pillow==8.4.0
protobuf==3.19.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
pyro-api==0.1.2
pyro-ppl==1.7.0
python-dateutil==2.8.2
pytz==2021.3
PyWavelets==1.1.1
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
scikit-image==0.18.3
scikit-learn==1.0
scipy==1.7.1
six==1.16.0
sklearn==0.0
statsmodels==0.13.0
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
threadpoolctl==3.0.0
tifffile==2021.10.12
tikzplotlib==0.9.15
torch==1.10.0
torchvision==0.11.1
tqdm==4.62.3
typing-extensions==3.10.0.2
uncertainty-metrics==0.0.81
urllib3==1.26.7
Werkzeug==2.0.2
